pyproject.toml
[tool.poetry]
name = "assistants-examples"
version = "0.1.0"
description = ""
authors = ["Caleb <cmwetherell@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.30.1"
phidata = "^2.4.10"
duckduckgo-search = "^6.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
